menyelesaikan api RSonline sesuai dengan masukan-masukan dari teman2

+++ resources/views/roles/index.blade.php
@@ -9,9 +9,9 @@
                     <div class="card">
                         <div class="card-header">
                             {{-- <h2 class="card-title">Role Management</h2> --}}
-                            {{-- @can('role-create') --}}
-                            <a class="btn btn-primary btn-sm" href="{{ route('roles.create') }}">Buat Grub Akses</a>
-                            {{-- @endcan --}}
+                            @can('role-create')
+                                <a class="btn btn-primary btn-sm" href="{{ route('roles.create') }}">Buat Grub Akses</a>
+                            @endcan
                         </div>
                         <div class="card-body">
                             <table class="table table-bordered table-hover">
@@ -28,13 +28,15 @@
 
                                             {{-- <a class="btn btn-success btn-sm"
                                                 href="{{ route('roles.show', $role->id) }}">Lihat</a> --}}
-                                            <a class="btn btn-warning btn-sm"
-                                                href="{{ route('roles.edit', $role->id) }}">Edit</a>
-                                            {{-- @can('role-delete') --}}
-                                            {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' => 'display:inline']) !!}
-                                            {!! Form::submit('Hapus', ['class' => 'btn btn-danger btn-sm']) !!}
-                                            {!! Form::close() !!}
-                                            {{-- @endcan --}}
+                                            @can('role-edit')
+                                                <a class="btn btn-warning btn-sm"
+                                                    href="{{ route('roles.edit', $role->id) }}">Edit</a>
+                                            @endcan
+                                            @can('role-delete')
+                                                {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' => 'display:inline']) !!}
+                                                {!! Form::submit('Hapus', ['class' => 'btn btn-danger btn-sm']) !!}
+                                                {!! Form::close() !!}
+                                            @endcan
 
                                         </td>
                                     </tr>
